+++ HWIOAuthBundle.php
@@ -12,7 +12,6 @@
 namespace HWI\Bundle\OAuthBundle;
 
 use HWI\Bundle\OAuthBundle\DependencyInjection\CompilerPass\ResourceOwnerMapCompilerPass;
-use HWI\Bundle\OAuthBundle\DependencyInjection\CompilerPass\SetResourceOwnerServiceNameCompilerPass;
 use HWI\Bundle\OAuthBundle\DependencyInjection\HWIOAuthExtension;
 use HWI\Bundle\OAuthBundle\DependencyInjection\Security\Factory\OAuthAuthenticatorFactory;
 use HWI\Bundle\OAuthBundle\DependencyInjection\Security\Factory\OAuthFactory;
@@ -45,7 +44,6 @@ class HWIOAuthBundle extends Bundle
             $extension->addSecurityListenerFactory(new OAuthFactory());
         }
 
-        $container->addCompilerPass(new SetResourceOwnerServiceNameCompilerPass());
         $container->addCompilerPass(new ResourceOwnerMapCompilerPass());
     }
 
@@ -55,11 +53,7 @@ class HWIOAuthBundle extends Bundle
     public function getContainerExtension()
     {
         // return the right extension instead of "auto-registering" it. Now the
-        // alias can be hwi_oauth instead of hwi_o_auth..
-        if (null === $this->extension) {
-            return new HWIOAuthExtension();
-        }
-
-        return $this->extension;
+        // alias can be hwi_oauth instead of hwi_o_auth.
+        return $this->extension ?: new HWIOAuthExtension();
     }
 }