+++ Tests/OAuth/ResourceOwner/GenericOAuth2ResourceOwnerTest.php
@@ -327,8 +327,6 @@ json;
     public function testGetSetName()
     {
         $this->assertEquals($this->resourceOwnerName, $this->resourceOwner->getName());
-        $this->resourceOwner->setName('foo');
-        $this->assertEquals('foo', $this->resourceOwner->getName());
     }
 
     public function testCsrfTokenIsValidWhenDisabled()