+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 Changelog
 =========
 ## 2.0.0 (2021-xx-xx)
-* Dropped PHP 7.3 support,
+* BC Break: Dropped PHP 7.3 support,
 * BC Break: `OAuthExtension` is now a lazy Twig extension using a Runtime,
 * BC Break: removed support for `FOSUserBundle`,
 * BC Break: changed `process()` argument for `Form/RegistrationFormHandlerInterface`, from `Form $form` to `FormInterface $form`,
@@ -13,6 +13,9 @@ Changelog
 * BC Break: changed `ResourceOwnerMapInterface::getResourceOwnerByName` signature, update if you use a custom resource owner,
 * BC Break: changed `ResourceOwnerMapInterface::getResourceOwnerByRequest` signature, update if you use a custom resource owner,
 * BC Break: changed `ResourceOwnerMapInterface::getResourceOwnerCheckPath` signature, update if you use a custom resource owner,
+* BC Break: `ResourceOwnerMap` uses service locator instead of DI container,
+* BC Break: Removed abstract services: `hwi_oauth.abstract_resource_owner.generic`, `hwi_oauth.abstract_resource_owner.oauth1` & `hwi_oauth.abstract_resource_owner.oauth2`,
+* BC Break: Removed `setName()` method from `OAuth/ResourceOwnerInterface`,
 
 ## 1.4.1 (2021-07-28)
 * Bugfix: Define missing `hwi_oauth.connect.confirmation` parameter,